Update README.md
This case study presents the analysis of a clinical trial conducted by Pymaceuticals, Inc., focusing on potential treatments for squamous cell carcinoma (SCC), a common form of skin cancer. As a senior data analyst, the objective is to visualize and summarize the performance of various drug regimens, particularly the company's drug of interest, Capomulin, using a dataset from an animal study involving 249 mice.